scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Scripts to scrape the mangas to show on the website"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fuzzywuzzy>=0.18.0",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "playwright-stealth>=2.0.0",
    "polars>=1.34.0",
    "pyarrow>=22.0.0",
    "pytest-playwright>=0.7.1",
    "python-levenshtein>=0.27.3",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
]
